requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
PyQt5==5.15.10
PyQt5-Qt5==5.15.12
PyQt5-sip==12.13.0
PyQt5-stubs==5.15.6.0
python-dotenv==1.0.1
requests==2.31.0
urllib3==2.2.1